fix(tests): move shared test mixins to tests.common to fix DeviceContext leak
pt_expt tests imported TestCaseSingleFrameWithNlist and get_tols from
tests.pt.model, which triggered tests/pt/__init__.py's
torch.set_default_device("cuda:9999999"). This pushed a DeviceContext
onto the mode stack, causing torch.zeros() (without device=) to target
a fake CUDA device — crashing on CPU-only machines.

Fix: move the shared mixins (pure numpy, no torch dependency) to
tests/common/test_mixins.py. The pt tests re-export from there for
backward compat. pt_expt tests now import directly from tests.common,
avoiding the tests.pt package entirely.

Also simplify the pt_expt conftest.py and remove the manual
_pop_device_contexts() call from test_change_bias.py.
