+++ source/tests/pt/model/test_env_mat.py
@@ -21,58 +21,9 @@ from ...seed import (
 dtype = env.GLOBAL_PT_FLOAT_PRECISION
 
 
-class TestCaseSingleFrameWithNlist:
-    def setUp(self) -> None:
-        # nloc == 3, nall == 4
-        self.nloc = 3
-        self.nall = 4
-        self.nf, self.nt = 2, 2
-        self.coord_ext = np.array(
-            [
-                [0, 0, 0],
-                [0, 1, 0],
-                [0, 0, 1],
-                [0, -2, 0],
-            ],
-            dtype=np.float64,
-        ).reshape([1, self.nall, 3])
-        self.atype_ext = np.array([0, 0, 1, 0], dtype=int).reshape([1, self.nall])
-        self.mapping = np.array([0, 1, 2, 0], dtype=int).reshape([1, self.nall])
-        # sel = [5, 2]
-        self.sel = [5, 2]
-        self.sel_mix = [7]
-        self.natoms = [3, 3, 2, 1]
-        self.nlist = np.array(
-            [
-                [1, 3, -1, -1, -1, 2, -1],
-                [0, -1, -1, -1, -1, 2, -1],
-                [0, 1, -1, -1, -1, -1, -1],
-            ],
-            dtype=int,
-        ).reshape([1, self.nloc, sum(self.sel)])
-        self.rcut = 2.2
-        self.rcut_smth = 0.4
-        # permutations
-        self.perm = np.array([2, 0, 1, 3], dtype=np.int32)
-        inv_perm = np.array([1, 2, 0, 3], dtype=np.int32)
-        # permute the coord and atype
-        self.coord_ext = np.concatenate(
-            [self.coord_ext, self.coord_ext[:, self.perm, :]], axis=0
-        ).reshape(self.nf, self.nall * 3)
-        self.atype_ext = np.concatenate(
-            [self.atype_ext, self.atype_ext[:, self.perm]], axis=0
-        )
-        self.mapping = np.concatenate(
-            [self.mapping, self.mapping[:, self.perm]], axis=0
-        )
-
-        # permute the nlist
-        nlist1 = self.nlist[:, self.perm[: self.nloc], :]
-        mask = nlist1 == -1
-        nlist1 = inv_perm[nlist1]
-        nlist1 = np.where(mask, -1, nlist1)
-        self.nlist = np.concatenate([self.nlist, nlist1], axis=0)
-        self.atol = 1e-12
+from ...common.test_mixins import (
+    TestCaseSingleFrameWithNlist,
+)
 
 
 class TestCaseSingleFrameWithNlistWithVirtual:
