+++ source/tests/pt_expt/conftest.py
@@ -1,28 +1,14 @@
 # SPDX-License-Identifier: LGPL-3.0-or-later
 """Conftest for pt_expt tests.
 
-Clears any leaked ``torch.utils._device.DeviceContext`` modes that may
-have been left on the torch function mode stack by ``make_fx`` or other
-tracing utilities during test collection.  A stale ``DeviceContext``
-silently reroutes ``torch.tensor(...)`` calls (without an explicit
-``device=``) to a fake CUDA device, causing spurious "no NVIDIA driver"
-errors on CPU-only machines.
+Safety net: pops any leaked ``torch.utils._device.DeviceContext`` modes
+from the torch function mode stack before each test.
 
-The leak is triggered when pytest collects descriptor test modules that
-import ``make_fx``.  A ``DeviceContext(cuda:127)`` ends up on the
-``torch.overrides`` function mode stack and is never popped.
-
-Our own code (``display_if_exist`` in ``deepmd/dpmodel/loss/loss.py``)
-is already fixed to pass ``device=`` explicitly.  However, PyTorch's
-``Adam._init_group`` (``torch/optim/adam.py``) contains::
-
-    torch.tensor(0.0, dtype=_get_scalar_dtype())  # no device=
-
-on the ``capturable=False, fused=False`` path (the default).  This is
-a PyTorch bug — the ``capturable=True`` branch correctly uses
-``device=p.device`` but the default branch omits it.  We cannot fix
-PyTorch internals, so this fixture works around the issue by popping
-leaked ``DeviceContext`` modes before each test.
+The primary leak source was ``source/tests/pt/__init__.py`` which calls
+``torch.set_default_device("cuda:9999999")``; pt_expt tests previously
+imported shared mixins from ``tests.pt.model``, triggering that init.
+This was fixed by moving the shared mixins to ``tests.common.test_mixins``
+so pt_expt tests no longer import from the ``tests.pt`` package.
 """
 
 import pytest
@@ -48,22 +34,8 @@ def _pop_device_contexts() -> list:
     return popped
 
 
-@pytest.fixture(autouse=True, scope="session")
-def _clear_leaked_device_context_session():
-    """Pop any stale DeviceContext once at session start.
-
-    This runs before any setUpClass, preventing CUDA init errors
-    in tests that call trainer.run() during class setup.
-    """
-    _pop_device_contexts()
-    yield
-
-
 @pytest.fixture(autouse=True)
 def _clear_leaked_device_context():
-    """Pop any stale ``DeviceContext`` before each test, restore after."""
-    popped = _pop_device_contexts()
+    """Pop any stale ``DeviceContext`` before each test (safety net)."""
+    _pop_device_contexts()
     yield
-    # Restore in reverse order so the stack is back to its original state.
-    for ctx in reversed(popped):
-        ctx.__enter__()
