+++ source/tests/pt_expt/descriptor/test_se_e2_a.py
@@ -21,10 +21,8 @@ from deepmd.pt_expt.utils.exclude_mask import (
     PairExcludeMask,
 )
 
-from ...pt.model.test_env_mat import (
+from ...common.test_mixins import (
     TestCaseSingleFrameWithNlist,
-)
-from ...pt.model.test_mlp import (
     get_tols,
 )
 from ...seed import (
