+++ source/tests/pt_expt/fitting/test_property_fitting.py
@@ -17,7 +17,7 @@ from deepmd.pt_expt.utils import (
     env,
 )
 
-from ...pt.model.test_env_mat import (
+from ...common.test_mixins import (
     TestCaseSingleFrameWithNlist,
 )
 from ...seed import (
