+++ source/tests/pt_expt/utils/test_exclusion_mask.py
@@ -12,7 +12,7 @@ from deepmd.pt_expt.utils.exclude_mask import (
     PairExcludeMask,
 )
 
-from ...pt.model.test_env_mat import (
+from ...common.test_mixins import (
     TestCaseSingleFrameWithNlist,
 )
